+++ docs/releases.rst
@@ -13,7 +13,9 @@ Version 1.4.0
 ~~~~~~~~~~~~~
 
 - Tagged with ``v1.4.0``.
-- Added support for JSON encoded environment values.
+- Added support for JSON encoded environment values if decorated with
+  ``:jsonenv``.  If JSON decoding fails (invalid JSON string), the value
+  is used as is.
 - Added a ``CODE-OF-CONDUCT`` file for contributors.
 
 .. _releases-1.3.0: